[backend]refact pagination
