+++ backend/src/main/scala/very/util/entity/Pagination.scala
@@ -1,0 +1,9 @@
+package very.util.entity
+
+case class Pagination(page: Int, pageSize: Int) {
+  assert(pageSize <= 50)
+  assert(page > 0)
+  def offset: Int = (page - 1) * pageSize
+  def limit: Int = pageSize
+
+}
