+++ backend/src/main/scala/very/util/web/PaginationSupport.scala
@@ -1,14 +1,6 @@
 package very.util.web
 
-import com.timzaak.fornet.dao.DB
-
-case class Pagination(page: Int, pageSize: Int) {
-  assert(pageSize <= 50)
-  assert(page > 0)
-  def offset: Int = (page - 1) * pageSize
-  def limit: Int = pageSize
-
-}
+import very.util.entity.Pagination
 
 trait PaginationSupport { this: org.scalatra.ScalatraBase =>
 
@@ -19,12 +11,12 @@ trait PaginationSupport { this: org.scalatra.ScalatraBase =>
   given pagination: Pagination = Pagination(page, pageSize)
 
   inline def search[T](
-                        arguments: Map[String, String => T => Boolean]
-                      ): Iterable[T => Boolean] = {
+    arguments: Map[String, String => T => Boolean]
+  ): Iterable[T => Boolean] = {
     for {
       (k, func) <- arguments
       value <- params.get(k) if value.nonEmpty
     } yield func(value)
-    
+
   }
 }
